requirements.txt
numpy
streamlit==1.29.0
tensorflow==2.19.0
Pillow==10.0.0
protobuf==3.20.3
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
scikit-learn==1.3.0
plotly==5.15.0
watchdog==3.0.0
opencv-python-headless==4.8.0.76
